+++ src/ezdxf/math/bspline.py
@@ -223,8 +223,8 @@ rather than normal 3D coordinates.
 from typing import List, Iterable, Sequence, TYPE_CHECKING, Dict, Tuple, Optional
 import math
 from math import pow, isclose
-from .vector import Vector, distance, NULLVEC
-from .linalg import Matrix
+from .vector import Vector, distance
+from .linalg import LUDecomposition, Matrix
 from ezdxf.lldxf.const import DXFValueError
 
 if TYPE_CHECKING:
@@ -542,8 +542,8 @@ def global_curve_interpolation(fit_points: Sequence['Vertex'],
 
     knots = list(control_frame_knots(len(fit_points) - 1, degree, t_vector))
     spline = Basis(knots=knots, order=degree + 1, count=len(fit_points))
-    matrix = Matrix([spline.basis(t) for t in t_vector])
-    control_points = matrix.gauss_matrix_solver([list(row) for row in fit_points])
+    lu = LUDecomposition([spline.basis(t) for t in t_vector])
+    control_points = lu.solve_matrix([list(row) for row in fit_points])
     return Vector.list(control_points.rows()), knots
 
 
@@ -571,9 +571,8 @@ def global_curve_interpolation_with_tangents(
 
     fit_points.insert(1, Vector(start_tangent) * knots[p + 1] / p)
     fit_points.insert(-1, Vector(end_tangent) * (1.0 - knots[m - p - 1]) / p)
-    matrix = Matrix(rows)
 
-    control_points = matrix.gauss_matrix_solver(fit_points)
+    control_points = LUDecomposition(rows).solve_matrix(fit_points)
     return Vector.list(control_points.rows()), knots
 
 
@@ -604,17 +603,17 @@ def global_curve_approximation(fit_points: Iterable['Vertex'],
     dn = fit_points[n]
     spline = Basis(knots, order=degree + 1, count=len(fit_points))
     # matrix_N[0] == row 0
-    matrix_N = [spline.basis(t) for t in t_vector]  # 0 .. n
+    N = [spline.basis(t) for t in t_vector]  # 0 .. n
 
-    def Q(k):
-        ntk = matrix_N[k]
+    def get_Q(k):
+        ntk = N[k]
         return fit_points[k] - d0 * ntk[0] - dn * ntk[h]
 
-    # matrix_Q[0] == row 1
-    matrix_Q = [sum(Q(k) * matrix_N[k][i] for k in range(1, n)) for i in range(1, h)]
-    matrix_N = Matrix([row[1:h] for row in matrix_N[1:-1]])
-    matrix_M = matrix_N.transpose() * matrix_N
-    P = matrix_M.gauss_matrix_solver(matrix_Q)
+    # Q[0] == row 1
+    Q = [sum(get_Q(k) * N[k][i] for k in range(1, n)) for i in range(1, h)]
+    N = Matrix([row[1:h] for row in N[1:-1]])
+    M = N.transpose() * N
+    P = LUDecomposition(M).solve_matrix(Q)
     control_points = [d0]
     control_points.extend(Vector.generate(P.rows()))
     control_points.append(dn)